+++ lib/collaboration_web/channels/public_channel.ex
@@ -1,7 +1,5 @@
 defmodule CollaborationWeb.PublicChannel do
   use CollaborationWeb, :channel
 
-  def join("public", _params, socket) do
-    {:ok, %{}, socket}
-  end
+  def join("public", _params, socket), do: {:ok, %{}, socket}
 end