+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.1
  * Added automated Prettier Codestyle for JS and CSS
  * If new ideas are posted in a topic that is not the currently active topic, a badge appears that indicates the number of new ideas in that topic. resets when topic is opened.
+ * My-Rating is now preserved in view when navigating through ideas / on page load.
 
 ## 2.0.0
 Complete Remade of the Application. Working Features: